zFe un-slop phase 1

+++ src/Speed/Indep/Src/FEng/FEngStandard.h
@@ -5,12 +5,14 @@
 #pragma once
 #endif
 
-void* FEngMalloc(unsigned int size, const char* pFilename, int Line);
-void FEngMemCpy(void* pDest, const void* pSrc, int Length);
-void FEngMemSet(void* pDest, int Value, int Length);
-void FEngStrCpy(char* pDest, const char* pSrc);
-int FEngStrLen(const char* pSrc);
-int FEngStrICmp(const char* pStr1, const char* pStr2);
+#include <cstddef>
+
+void *FEngMalloc(size_t size, const char *pFilename, int Line);
+void FEngMemCpy(void *pDest, const void *pSrc, int Length);
+void FEngMemSet(void *pDest, int Value, int Length);
+void FEngStrCpy(char *pDest, const char *pSrc);
+int FEngStrLen(const char *pSrc);
+int FEngStrICmp(const char *pStr1, const char *pStr2);
 void InitFEngMemoryPool();
 float FEngAbs(float x);
 float FEngSqrt(float x);
@@ -19,14 +21,14 @@ float FEngACos(float x);
 
 struct DummyFEngNewType {};
 
-inline void* operator new(unsigned int size, const char* file, int line, DummyFEngNewType* dummy) {
+inline void *operator new(size_t size, const char *file, int line, DummyFEngNewType *dummy) {
     return FEngMalloc(size, file, line);
 }
 
-inline void* operator new[](unsigned int size, const char* file, int line, DummyFEngNewType* dummy) {
+inline void *operator new[](size_t size, const char *file, int line, DummyFEngNewType *dummy) {
     return FEngMalloc(size, file, line);
 }
 
-#define FENG_NEW new(nullptr, 0, static_cast<DummyFEngNewType*>(nullptr))
+#define FENG_NEW new (nullptr, 0, static_cast<DummyFEngNewType *>(nullptr))
 
 #endif
